Change to using rRESPA integrator in main
If no 'RESPA' or 'respa' keyword is specified in the CONF input,
nothing is changed and the rRESAP scheme reduces automatically to a
normal Velocity Verlet integrator.

If no molecular forces are present (no 'molecules' is found in the
input .hdf5), the rRESPA integrator reduces automatically to a
(inefficient) Velocity Verlet integrator using only the field forces.