Fix bug causing angle force to sometimes become inf
When rij and rkj are very close to parallel, the denominator in the
force expression diverges as 1/sin(theta). The numerator vanishes
sufficiently rapidly to cancel the divergence, but the numerical
expression is unstable. In the cases where cos(theta) is Nπ to
numerical precision, we just ignore the force and the energy. Mirrors
what is done in GROMACS.