Clean COM velocities before thermostatting
Removes the center of mass momentum before applying the thermostat,
thus performing the velocity rescaling on the correct number of degrees
of freedom.

Also change the implementation of the CSVR thermostat to align with the
derivation from the 2008 Comput. Phys. Commun paper, not the original
2007 J. Chem. Phys. paper. Change call signature of csvr_thermostat to
change velocities in place, returning None.