Update field.py
volume per cell - leave for now, but remember that multipltying might be and option. Note: I check for lipid-A w. pressure: for PE the max value on the force mesh is 9.0 smt, and for GPE it is 7.78  ish w/o multiplying, multiplying with volume_per_cell gives 0.061326362, which seems a bit small to me. For now I change it back.